Time: 0 ms (100.00%) | Memory: 9 MB (73.32%) - LeetSync

+++ 28-find-the-index-of-the-first-occurrence-in-a-string/find-the-index-of-the-first-occurrence-in-a-string.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int strStr(string haystack, string needle) {
-         int pos = haystack.find(needle);
+         size_t pos = haystack.find(needle);
 
          if(pos != string::npos)
          return pos;
